Fix case of RefreshList returning empty player
There is a weird case, where RefreshList would return a player element
with empty UniqueId and Username. This can not be PavTV, as this
happened on an EFP server, which is not supported by PavTV, and it
should also not be an RCON client, since they have never showed up
in RefreshList. Very strange behaviour, as expected from Pavlov...